client app bugs fixing finalized

+++ mobile-app/client mobile/lib/models/Reservation.dart
@@ -1,14 +1,14 @@
 // ignore_for_file: file_names
 
-class Reservation{
+class Reservation {
   String client;
   String reservationPlateNumber;
   String branch;
   String branchName;
   int price;
   int startingTime;
-  int duration;
+  double duration;
 
-  Reservation(this.client, this.reservationPlateNumber, this.branch, this.branchName,
-      this.price, this.startingTime, this.duration);
+  Reservation(this.client, this.reservationPlateNumber, this.branch,
+      this.branchName, this.price, this.startingTime, this.duration);
 }
